refactor(commands): extract /sandbox-toggle command into sandbox.rs
